Move IsTraversal to the internal module.
