+++ fresnel/src/Fresnel/Traversal/Internal.hs
@@ -1,2 +1,15 @@
 module Fresnel.Traversal.Internal
-() where
+( IsTraversal
+) where
+
+import Control.Arrow (Kleisli)
+import Data.Profunctor (Forget, Star)
+import Data.Profunctor.Traversing (Traversing)
+import Fresnel.Optional.Internal (IsOptional)
+
+class (IsOptional p, Traversing p) => IsTraversal p
+
+instance IsTraversal (->)
+instance Monad m => IsTraversal (Kleisli m)
+instance Monoid r => IsTraversal (Forget r)
+instance Applicative f => IsTraversal (Star f)
